Updated to use yield and IEnumerable and cleaned up code

+++ TextControlBox/Text/UndoRedo.cs
@@ -2,7 +2,11 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
+using System.IO.Compression;
+using System.Text;
 using TextControlBox.Helper;
+using static System.Collections.Specialized.BitVector32;
 
 namespace TextControlBox.Text
 {
@@ -35,8 +39,22 @@ namespace TextControlBox.Text
             });
         }
 
+        private void RecordSingleLine(Action action, PooledList<Line> TotalLines, int startline)
+        {
+            var lineBefore = ListHelper.GetLine(TotalLines, startline).Content;
+            action.Invoke();
+            var lineAfter = ListHelper.GetLine(TotalLines, startline).Content;
+            AddUndoItem(null, startline, lineBefore, lineAfter, 1, 1);
+        }
+
         public void RecordUndoAction(Action action, PooledList<Line> TotalLines, int startline, int undocount, int redoCount, string NewLineCharacter)
         {
+            if (undocount == redoCount && redoCount == 1)
+            {
+                RecordSingleLine(action, TotalLines, startline);
+                return;
+            }
+
             var linesBefore = ListHelper.GetLinesAsString(TotalLines, startline, undocount, NewLineCharacter);
             action.Invoke();
             var linesAfter = ListHelper.GetLinesAsString(TotalLines, startline, redoCount, NewLineCharacter);
@@ -46,15 +64,19 @@ namespace TextControlBox.Text
         public void RecordUndoAction(Action action, PooledList<Line> TotalLines, TextSelection selection, int NumberOfAddedLines, string NewLineCharacter)
         {
             var orderedSel = Selection.OrderTextSelection(selection);
+            if (orderedSel.StartPosition.LineNumber == orderedSel.EndPosition.LineNumber && orderedSel.StartPosition.LineNumber == 1)
+            {
+                RecordSingleLine(action, TotalLines, orderedSel.StartPosition.LineNumber);
+                return;
+            }
 
-            //Check whether the selected lines are selected from start to end
-            if (!(orderedSel.EndPosition.CharacterPosition == ListHelper.GetLine(TotalLines, orderedSel.EndPosition.LineNumber).Length &&
-                orderedSel.StartPosition.CharacterPosition == 0) || Selection.WholeTextSelected(selection, TotalLines))
+            int NumberOfRemovedLines = orderedSel.EndPosition.LineNumber - orderedSel.StartPosition.LineNumber + 1;
+            if (NumberOfAddedLines == 0 && !Selection.WholeLinesAreSelected(selection, TotalLines))
             {
                 NumberOfAddedLines += 1;
             }
 
-            var linesBefore = ListHelper.GetLinesAsString(TotalLines, orderedSel.StartPosition.LineNumber, orderedSel.EndPosition.LineNumber - orderedSel.StartPosition.LineNumber + 1, NewLineCharacter);
+            var linesBefore = ListHelper.GetLinesAsString(TotalLines, orderedSel.StartPosition.LineNumber, NumberOfRemovedLines, NewLineCharacter);
             action.Invoke();
             var linesAfter = ListHelper.GetLinesAsString(TotalLines, orderedSel.StartPosition.LineNumber, NumberOfAddedLines, NewLineCharacter);
 
@@ -63,13 +85,13 @@ namespace TextControlBox.Text
                 orderedSel.StartPosition.LineNumber,
                 linesBefore,
                 linesAfter,
-                orderedSel.EndPosition.LineNumber - orderedSel.StartPosition.LineNumber + 1,
+                NumberOfRemovedLines,
                 NumberOfAddedLines
                 );
         }
 
         /// <summary>
-        /// Excecutes the undo and apply the changes to the text
+        /// Excecutes the undo and applys the changes to the text
         /// </summary>
         /// <param name="TotalLines">A list containing all the lines of the textbox</param>
         /// <param name="NewLineCharacter">The current line-ending character either CR, LF or CRLF</param>
@@ -88,10 +110,19 @@ namespace TextControlBox.Text
             UndoRedoItem item = UndoStack.Pop();
             RecordRedo(item);
 
+            //Faster for singleline
+            if (item.UndoCount == 1 && item.RedoCount == 1)
+            {
+                ListHelper.GetLine(TotalLines, item.StartLine).Content = StringManager.CleanUpString(item.UndoText);
+            }
+            else
+            {
+                ListHelper.RemoveRange(TotalLines, item.StartLine, item.RedoCount);
+                if (item.UndoCount > 0)
+                    ListHelper.InsertRange(TotalLines, ListHelper.GetLinesFromString(StringManager.CleanUpString(item.UndoText), NewLineCharacter), item.StartLine);
 
-            ListHelper.RemoveRange(TotalLines, item.StartLine, item.RedoCount);
-            if (item.UndoCount > 0)
-                ListHelper.InsertRange(TotalLines, ListHelper.GetLinesFromString(StringManager.CleanUpString(item.UndoText), NewLineCharacter), item.StartLine);
+                //Selection.ReplaceLines(TotalLines, item.StartLine, item.RedoCount, StringManager.CleanUpString(Decompress(item.UndoText)).Split(NewLineCharacter));
+            }
 
             return item.Selection;
         }
@@ -111,10 +142,19 @@ namespace TextControlBox.Text
             RecordUndo(item);
             HasRedone = true;
 
-            ListHelper.RemoveRange(TotalLines, item.StartLine, item.UndoCount);    
-            if (item.RedoCount > 0)
-                ListHelper.InsertRange(TotalLines, ListHelper.GetLinesFromString(StringManager.CleanUpString(item.RedoText), NewLineCharacter), item.StartLine);
+            //Faster for singleline
+            if (item.UndoCount == 1 && item.RedoCount == 1)
+            {
+                ListHelper.GetLine(TotalLines, item.StartLine).Content = StringManager.CleanUpString(item.RedoText);
+            }
+            else
+            {
+                ListHelper.RemoveRange(TotalLines, item.StartLine, item.UndoCount);
+                if (item.RedoCount > 0)
+                    ListHelper.InsertRange(TotalLines, ListHelper.GetLinesFromString(StringManager.CleanUpString(item.RedoText), NewLineCharacter), item.StartLine);
 
+                //Selection.ReplaceLines(TotalLines, item.StartLine, item.UndoCount, StringManager.CleanUpString(Decompress(item.RedoText)).Split(NewLineCharacter));
+            }
             return null;
         }
 
@@ -145,7 +185,6 @@ namespace TextControlBox.Text
         /// </summary>
         public bool CanRedo { get => RedoStack.Count > 0; }
     }
-
     public struct UndoRedoItem
     {
         public int StartLine { get; set; }
